requirements.txt
streamlit
yfinance
plotly
pandas
openai
python-dotenv
